Refactor statement validation tests
Refs #3741

* Move statement validation tests in parse/ into statement/
  * Combine with existing suites where applicable
